# backend/src/auth/auth.controller.ts
import {
  Controller,
  Post,
  Delete,
  Body,
  Res,
  Get,
  HttpCode,
  UseGuards,
  Request,
  UnauthorizedException,
  Patch,
} from '@nestjs/common';
import { Throttle } from '@nestjs/throttler';
import type { Response, Request as ExpressRequest } from 'express';
import { AuthService } from './auth.service';
import { RegisterDto } from './dto/register.dto';
import { LoginDto } from './dto/login.dto';
import { VerifyOtpDto } from './dto/verify-otp.dto';
import { ResendOtpDto } from './dto/resend-otp.dto';
import { JwtAuthGuard } from './guards/jwt-auth.guard';
import { ForgotPasswordDto } from './dto/forgot-password.dto';
import { UpdateProfileDto } from './dto/update-profile.dto';
import { ChangePasswordDto } from './dto/change-password.dto';

interface AuthenticatedRequest extends ExpressRequest {
  user: {
    id: string;
    email?: string;
  };
}

interface RequestWithCookies extends ExpressRequest {
  cookies: {
    refresh_token?: string;
  };
}

@Controller('auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  @Throttle({ default: { limit: 5, ttl: 60000 } })
  @Post('register')
  async register(@Body() dto: RegisterDto) {
    return this.authService.register(dto);
  }

  private getCookieOptions(maxAge: number): {
    httpOnly: boolean;
    secure: boolean;
    sameSite: 'none' | 'lax';
    maxAge: number;
  } {
    const isInProduct = process.env.NODE_ENV === 'production';
    return {
      httpOnly: true,
      secure: isInProduct,
      sameSite: isInProduct ? 'none' : 'lax',
      maxAge,
    };
  }

  @Throttle({ default: { limit: 5, ttl: 60000 } })
  @Post('verify-email')
  @HttpCode(200)
  async verifyEmail(
    @Body() dto: VerifyOtpDto,
    @Res({ passthrough: true }) res: Response,
  ) {
    const tokens = await this.authService.verifyEmail(dto.email, dto.code);

    res.cookie(
      'access_token',
      tokens.tokens.accessToken,
      this.getCookieOptions(15 * 60 * 1000),
    );

    res.cookie(
      'refresh_token',
      tokens.tokens.refreshToken,
      this.getCookieOptions(7 * 24 * 60 * 60 * 1000),
    );

    return { message: 'Email verified successfully.', user: tokens.user };
  }

  @Throttle({ default: { limit: 10, ttl: 60000 } })
  @Post('login')
  @HttpCode(200)
  async login(
    @Body() dto: LoginDto,
    @Res({ passthrough: true }) res: Response,
  ) {
    const result = await this.authService.login(dto);

    res.cookie(
      'access_token',
      result.accessToken,
      this.getCookieOptions(15 * 60 * 1000),
    );

    res.cookie(
      'refresh_token',
      result.refreshToken,
      this.getCookieOptions(7 * 24 * 60 * 60 * 1000),
    );

    return {
      message: 'Login successful.',
      user: result.user,
    };
  }

  @Get('universities')
  @HttpCode(200)
  async getUniversities() {
    return this.authService.getUniversities();
  }

  @Throttle({ default: { limit: 4, ttl: 60000 } })
  @Post('resend-otp')
  @HttpCode(200)
  async resendOtp(@Body() dto: ResendOtpDto) {
    return this.authService.resendOtp(dto.email);
  }

  @Post('logout')
  @HttpCode(200)
  logout(@Res({ passthrough: true }) res: Response) {
    res.clearCookie('access_token', this.getCookieOptions(0));
    res.clearCookie('refresh_token', this.getCookieOptions(0));
    return { message: 'Logged out successfully.' };
  }

  @UseGuards(JwtAuthGuard)
  @Get('me')
  async getMe(@Request() req: AuthenticatedRequest) {
    return this.authService.getMe(req.user.id);
  }

  @UseGuards(JwtAuthGuard)
  @Patch('me')
  @HttpCode(200)
  async updateMe(
    @Request() req: AuthenticatedRequest,
    @Body() dto: UpdateProfileDto,
  ) {
    return this.authService.updateProfile(req.user.id, dto);
  }

  @UseGuards(JwtAuthGuard)
  @Throttle({ default: { limit: 5, ttl: 60000 } })
  @Patch('me/password')
  @HttpCode(200)
  async changePassword(
    @Request() req: AuthenticatedRequest,
    @Body() dto: ChangePasswordDto,
  ) {
    return this.authService.changePassword(req.user.id, dto);
  }

  @UseGuards(JwtAuthGuard)
  @Delete('me')
  @HttpCode(200)
  async deleteMe(
    @Request() req: AuthenticatedRequest,
    @Res({ passthrough: true }) res: Response,
  ) {
    const result = await this.authService.deactivateAccount(req.user.id);

    res.clearCookie('access_token', this.getCookieOptions(0));
    res.clearCookie('refresh_token', this.getCookieOptions(0));

    return result;
  }

  @Throttle({ default: { limit: 4, ttl: 60000 } })
  @Post('forgot-password')
  @HttpCode(200)
  async forgotPassword(@Body() dto: ForgotPasswordDto) {
    return this.authService.forgotPassword(dto);
  }

  @Post('refresh')
  @HttpCode(200)
  async refresh(
    @Request() req: RequestWithCookies,
    @Res({ passthrough: true }) res: Response,
  ) {
    const refreshToken = req.cookies?.refresh_token;

    if (!refreshToken)
      throw new UnauthorizedException('No refresh token provided');

    const tokens = await this.authService.refreshTokens(refreshToken);

    res.cookie('access_token', tokens.accessToken, {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'lax' as const,
      maxAge: 15 * 60 * 1000,
    });

    res.cookie('refresh_token', tokens.refreshToken, {
      httpOnly: true,
      secure: process.env.NODE_ENV === 'production',
      sameSite: 'lax' as const,
      maxAge: 7 * 24 * 60 * 60 * 1000,
    });

    return { message: 'Token refreshed' };
  }
}