# docs/OpenAPI/openapi.yaml
openapi: 3.0.4
info:
  title: Uni Textbook Marketplace API
  description: >
    NexusDev - Uni Textbook Marketplace, COS 301 Software Engineering, 2026.
    Single source of truth for the API contract. Everyone edits this file
    directly, the same day and endpoint is committed. sas.tex's API Contracts
    section is a readbale rendering of this file, not an independently
    maintained copy. If the two disagree, this file is right.
  version: "5.0.0"
  contact:
    name: NexusDev
    url: https://github.com/COS301-SE-2026
servers:
  - url: https://nexusdev-backend.whitesand-df72b78b.southafricanorth.azurecontainerapps.io/api
    description: Production (main)
  - url: https://nexusdev-backend-staging.whitesand-df72b78b.southafricanorth.azurecontainerapps.io/api
    description: Staging (develop)
  - url: http://localhost:3001/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporterId:
          type: string
          format: uuid
          description: The ID of the user who submitted the report
        listingId:
          type: string
          format: uuid
          description: The ID of the listing being reported
        reason:
          type: string
          description: The reason for the report
        status:
          type: string
          enum: [PENDING, REVIEWED]
          description: The current status of the report
        createdAt:
          type: string
          format: date-time
          description: Timestamp of report creation
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not permitted (wrong role, not the owner)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: The request is invalid or could not be processed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /auth/register:
    post:
      operationId: registerStudent
      tags: [Auth]
      summary: Starts student registration. Checks the email domain and sends the OTP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, surname, university, email]
              properties:
                fullName:
                  type: string
                surname:
                  type: string
                university:
                  type: string
                email:
                  type: string
      responses:
        "201":
          description: Registration successful, OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  userId:
                    type: string
                    format: uuid
        "400":
          description: Missing required field, or email domain not in allowlist
        "409":
          description: Email address is already registered
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [Auth]
      summary: Verifies the 6-digit code sent to the student's email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                code:
                  type: string
                  minLength: 6
                  maxLength: 6
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  verified:
                    type: boolean
        "400":
          description: OTP is incorrect or has expired
        "404":
          description: No pending OTP found for this email
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Logs in a verified student and hands back a signed JWT (cookie).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid email or password
        "403":
          description: Account exists but email is not verified
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [Auth]
      summary: Resends the email verification code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        "200":
          description: Verification code resent
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [Auth]
      summary: Starts the password recovery flow.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Password recovery request accepted
        "400":
          description: Invalid request
  /auth/me:
    get:
      operationId: getMyProfile
      tags: [Auth]
      summary: Returns the logged-in user's profile.
      responses:
        "200":
          description: The current user's profile
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMyProfile
      tags: [Auth]
      summary: Updates the logged-in user's profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  minLength: 1
                  maxLength: 100
                last_name:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        "200":
          description: Profile updated
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deactivateMyAccount
      tags: [Auth]
      summary: Deactivates the logged-in user's account and clears session cookies.
      responses:
        "200":
          description: Account deactivated
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me/password:
    patch:
      operationId: changeMyPassword
      tags: [Auth]
      summary: Changes the logged-in user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password, confirm_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
                confirm_password:
                  type: string
      responses:
        "200":
          description: Password changed
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Clears the session cookie.
      responses:
        "200":
          description: Logged out
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/universities:
    get:
      operationId: getUniversities
      tags: [Auth]
      summary: Lists supported universities, for the registration dropdown.
      security: []
      responses:
        "200":
          description: List of universities
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Auth]
      summary: Gets a new access token using a refresh token.
      security: []
      responses:
        "200":
          description: New access token issued
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /admin/{id}/approve:
    patch:
      operationId: approveListing
      tags: [Admin]
      summary: Approves a pending listing. Sets status APPROVED, writes audit log.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Listing approved
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/{id}/reject:
    patch:
      operationId: rejectListing
      tags: [Admin]
      summary: Rejects a pending listing. Sets status REJECTED, writes audit log.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Listing rejected
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/audit-log:
    get:
      operationId: getAuditLog
      tags: [Admin]
      summary: Returns audit log entries for admin review.
      parameters:
        - in: query
          name: performedBy
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: entityType
          schema: { type: string }
        - in: query
          name: entityId
          schema: { type: string }
        - in: query
          name: startDate
          schema: { type: string }
        - in: query
          name: endDate
          schema: { type: string }
      responses:
        "200":
          description: Audit log entries
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/emails:
    get:
      operationId: getAdminEmails
      tags: [Admin]
      summary: Returns a list of admin emails, used for filters.
      responses:
        "200":
          description: Admin emails
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/reports:
    get:
      operationId: getReportsForAdmin
      tags: [Reports]
      summary: Returns reports for admin review, paginated, filterable by status. 
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, REVIEWED] }
        - in: query
          name: page
          schema: { type: integer }
          description: Page number
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Number of reports returned per page.
      responses:
        "200":
          description: List of reports for admin review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/reports/{id}:
    get:
      operationId: getReportById
      tags: [Reports]
      summary: Returns one report for admin review.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/reports/{id}/dismiss:
    patch:
      operationId: dismissReport
      tags: [Reports]
      summary: Dismisses a report without banning the reported user.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
          description: The ID of the report to dismiss
      responses:
        "200":
          description: Report dismissed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/{id}/ban:
    patch:
      operationId: banUser
      tags: [Reports]
      summary: >
        Bans the user who owns the reported listing. Writes an audit_log
        entry (action = BANNED) and triggers a ban-notice notification with
        appeal instructions via NotificationService.emit().
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for banning the user
      responses:
        "200":
          description: User banned
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
security:
  - bearerAuth: []